requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black @ git+git://github.com/psf/black@692c0f50d91e3163bb87401e4a0e070b2eb5b163
cfgv==3.2.0
click==7.1.2
distlib==0.3.1
Django==3.1.5
djangorestframework==3.12.2
filelock==3.0.12
identify==1.5.13
mypy-extensions==0.4.3
nodeenv==1.5.0
pathspec==0.8.1
pre-commit==2.9.3
pytz==2020.5
PyYAML==5.3.1
regex==2020.11.13
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
virtualenv==20.3.1
